# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  images: string[];
  shortDescription: string;
  /** @nullable */
  description?: string | null;
  features?: string[];
  colors?: string[];
  sizes?: string[];
  /** @nullable */
  warrantyYears?: number | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  inStock: boolean;
  isFeatured?: boolean;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  material?: string | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  image?: string | null;
  productCount: number;
  /** @nullable */
  description?: string | null;
}

export interface OrderItem {
  productId: number;
  quantity: number;
  price: number;
  name: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface OrderInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  address: string;
  city: string;
  state: string;
  pinCode: string;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  couponCode?: string | null;
  paymentMethod: string;
  /** @nullable */
  razorpayOrderId?: string | null;
}

export interface Order {
  id: number;
  orderId: string;
  status: string;
  totalAmount: number;
  fullName?: string;
  phone?: string;
  createdAt: string;
}

export interface WarrantyRegistrationInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  productName: string;
  purchaseDate: string;
  invoiceNumber: string;
  /** @nullable */
  dealerName?: string | null;
}

export interface WarrantyRegistration {
  id: number;
  registrationId: string;
  status: string;
  createdAt: string;
}

export interface WarrantyClaimInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  registrationId: string;
  issueDescription: string;
  /** @nullable */
  productName?: string | null;
}

export interface WarrantyClaim {
  id: number;
  claimId: string;
  status: string;
  createdAt: string;
}

export interface NewsletterInput {
  email: string;
  /** @nullable */
  name?: string | null;
}

export interface NewsletterSubscription {
  id: number;
  email: string;
  createdAt: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
sort?: ListProductsSort;
minPrice?: number;
maxPrice?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  newest: 'newest',
  popular: 'popular',
} as const;