# lib/api-zod/src/generated/types/newsletterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterInput {
  email: string;
  /** @nullable */
  name?: string | null;
}