# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  orderId: string;
  status: string;
  totalAmount: number;
  fullName?: string;
  phone?: string;
  createdAt: string;
}