# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  address: string;
  city: string;
  state: string;
  pinCode: string;
  items: OrderItem[];
  totalAmount: number;
  /** @nullable */
  couponCode?: string | null;
  paymentMethod: string;
  /** @nullable */
  razorpayOrderId?: string | null;
}