# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  quantity: number;
  price: number;
  name: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}