# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  images: string[];
  shortDescription: string;
  /** @nullable */
  description?: string | null;
  features?: string[];
  colors?: string[];
  sizes?: string[];
  /** @nullable */
  warrantyYears?: number | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  inStock: boolean;
  isFeatured?: boolean;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  material?: string | null;
}