# lib/api-zod/src/generated/types/warrantyClaimInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface WarrantyClaimInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  registrationId: string;
  issueDescription: string;
  /** @nullable */
  productName?: string | null;
}