# lib/api-zod/src/generated/types/warrantyRegistrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Yunora ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface WarrantyRegistrationInput {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  productName: string;
  purchaseDate: string;
  invoiceNumber: string;
  /** @nullable */
  dealerName?: string | null;
}